requirements.txt
tensorflow
scikit-learn
pandas
matplotlib
numpy
requests
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
google-auth
jupyter
notebook
python-dotenv

